Syncing files from `cloud/release/2023.009` to `release/2023.009`

+++ applications/vanilla/models/class.categorymodel.php
@@ -2,7 +2,7 @@
 /**
  * Category model
  *
- * @copyright 2009-2022 Vanilla Forums Inc.
+ * @copyright 2009-2023 Vanilla Forums Inc.
  * @license GPL-2.0-only
  * @package Vanilla
  * @since 2.0
@@ -95,6 +95,8 @@ class CategoryModel extends Gdn_Model implements
     public const NOTIFICATION_DISCUSSIONS = "discussions";
 
     public const NOTIFICATION_FOLLOW = "follow";
+
+    public const DEFAULT_PREFERENCE_DATE_FOLLOWED = "2023-04-18 23:59:59";
 
     /** Cache key. */
     const CACHE_KEY = "Categories";
@@ -402,7 +404,7 @@ class CategoryModel extends Gdn_Model implements
         $allowedDiscussionTypes = self::getAllowedDiscussionData($row);
         $allowedDiscussionTypes = array_keys($allowedDiscussionTypes);
 
-        $discussionTypes = array_intersect($allowedDiscussionTypes, $categoryAllowedDiscussionTypes);
+        $discussionTypes = array_intersect($categoryAllowedDiscussionTypes, $allowedDiscussionTypes);
 
         return $discussionTypes ?? [];
     }
@@ -530,11 +532,12 @@ class CategoryModel extends Gdn_Model implements
     /**
      * Get searchable category IDs.
      *
-     * @param int $categoryID The root category ID.
+     * @param int|null $categoryID The root category ID.
      * @param bool|null $followedCategories If set, include or exclude followed categories.
      * @param bool|null $includeChildCategories Get child category IDs as well.
      * @param bool|null $includeArchivedCategories If set include archived categories.
-     *
+     * @param array|null $categoryIDs
+     * @param string|null $categorySearch
      * @return int[] CategoryIDs.
      */
     public function getSearchCategoryIDs(
@@ -744,12 +747,9 @@ class CategoryModel extends Gdn_Model implements
                 throw new ClientException(t("Already following the maximum number of categories."));
             }
         }
+        $fields = $this->getInsertUpdateCategoryPreferenceFields($userID, $categoryID, $followed);
 
-        $this->SQL->replace(
-            "UserCategory",
-            ["Followed" => $followed],
-            ["UserID" => $userID, "CategoryID" => $categoryID]
-        );
+        $this->SQL->replace("UserCategory", $fields, ["UserID" => $userID, "CategoryID" => $categoryID]);
         static::clearUserCache();
         Gdn::cache()->remove("Follow_{$userID}");
 
@@ -757,6 +757,20 @@ class CategoryModel extends Gdn_Model implements
         return $result;
     }
 
+    /**
+     * Get insert update fields for category following
+     *
+     * @param int $userID
+     * @param int $categoryID
+     * @param int $followed
+     * @return int[]
+     */
+    private function getInsertUpdateCategoryPreferenceFields(int $userID, int $categoryID, int $followed): array
+    {
+        $fields = ["Followed" => $followed, "Unfollow" => $followed ? 0 : 1];
+        $fields[$followed ? "DateFollowed" : "DateUnfollowed"] = DateTimeFormatter::getCurrentDateTime();
+        return $fields;
+    }
     /**
      * Get the enabled status of category following, returned as a boolean value.
      *
@@ -837,6 +851,7 @@ class CategoryModel extends Gdn_Model implements
      *   - filterHideDiscussions (bool): Filter out categories with a truthy HideAllDiscussions column?
      *   - filterArchivedCategories (bool): Filter out categories that are archived.
      *   - forceArrayReturn (bool): Force an array return value.
+     *   - filterNonDiscussionCategories (bool) : Filter out categories with no discussion in them
      * @return array|bool An array of filtered categories or true if no categories were filtered.
      */
     public function getVisibleCategories(array $options = [])
@@ -861,6 +876,7 @@ class CategoryModel extends Gdn_Model implements
         $filterHideDiscussions = $options["filterHideDiscussions"] ?? false;
         $filterArchivedCategories = $options["filterArchivedCategories"] ?? false;
         $filterNonPostableCategories = $options["filterNonPostableCategories"] ?? false;
+        $filterNonDiscussionCategories = $options["filterNonDiscussionCategories"] ?? false;
 
         foreach ($categories as $categoryID => $category) {
             if ($filterHideDiscussions && ($category["HideAllDiscussions"] ?? false)) {
@@ -869,6 +885,11 @@ class CategoryModel extends Gdn_Model implements
             }
 
             if ($filterArchivedCategories && ($category["Archived"] ?? false)) {
+                $unfiltered = false;
+                continue;
+            }
+
+            if ($filterNonDiscussionCategories && $category["CountDiscussions"] == 0) {
                 $unfiltered = false;
                 continue;
             }
@@ -1477,6 +1498,7 @@ class CategoryModel extends Gdn_Model implements
                         ? $this->getCategoryAllowedDiscussionTypes($category)
                         : ["Discussion"];
                     $discussionTypes = array_map("lcfirst", $discussionTypes);
+                    $discussionTypes = array_values($discussionTypes);
                     $category["AllowedDiscussionTypes"] = $discussionTypes;
                     setValue($field, $row, $category);
                 }
@@ -2513,14 +2535,15 @@ class CategoryModel extends Gdn_Model implements
      *
      * @param array $categories
      * @param bool $addUserCategory
+     * @param int | null $userID
      */
-    public static function joinUserData(&$categories, $addUserCategory = true)
+    public static function joinUserData(array &$categories, bool $addUserCategory = true, ?int $userID = null)
     {
         $iDs = array_column($categories, "CategoryID", "CategoryID");
         $categories = array_combine($iDs, $categories);
 
         if ($addUserCategory) {
-            $userData = self::instance()->getUserCategories();
+            $userData = self::instance()->getUserCategories($userID);
 
             foreach ($iDs as $iD) {
                 $category = $categories[$iD];
@@ -2549,6 +2572,8 @@ class CategoryModel extends Gdn_Model implements
                 $categories[$iD]["Following"] = $following;
 
                 $categories[$iD]["Followed"] = boolval($row["Followed"] ?? false);
+
+                $categories[$iD]["DateFollowed"] = $row["DateFollowed"] ?? null;
 
                 // Calculate the read field.
                 if ($category["DisplayAs"] == self::DISPLAY_HEADING) {
@@ -2627,6 +2652,33 @@ class CategoryModel extends Gdn_Model implements
 
         if ($dbRecord["ParentCategoryID"] <= 0) {
             $dbRecord["ParentCategoryID"] = null;
+        }
+
+        if (ModelUtils::isExpandOption("lastPost", $expand) && !empty($dbRecord["LastDiscussionID"])) {
+            $recentPost = [];
+            $valid = true;
+            $mappings = [
+                "discussionID" => "LastDiscussionID",
+                "commentID" => "LastCommentID",
+                "name" => "LastTitle",
+                "url" => "LastUrl",
+                "dateInserted" => "LastDateInserted",
+                "insertUserID" => "LastUserID",
+            ];
+            foreach ($mappings as $key => $mapping) {
+                if (empty($dbRecord[$mapping]) && $mapping != "LastCommentID") {
+                    $valid = false;
+                    break;
+                }
+                $recentPost[$key] = $dbRecord[$mapping];
+            }
+            if ($valid) {
+                $dbRecord["lastPost"] = $recentPost;
+                if (!empty($dbRecord["LastUser"])) {
+                    $dbRecord["lastPost"]["insertUser"] = $dbRecord["LastUser"];
+                    unset($dbRecord["LastUser"]);
+                }
+            }
         }
 
         $dbRecord["Name"] = empty($dbRecord["Name"]) ? t("Untitled") : $dbRecord["Name"];
@@ -2719,7 +2771,7 @@ class CategoryModel extends Gdn_Model implements
                 yield;
             }
         }
-        $this->prepareForDelete($categoryID);
+        $this->prepareForDelete($categoryID, false);
         $this->deleteInternal($categoryID, true);
         return LongRunner::FINISHED;
     }
@@ -3210,7 +3262,6 @@ class CategoryModel extends Gdn_Model implements
             }
             unset($where["Followed"]);
         }
-
         $result = parent::getWhere($where, $orderFields, $orderDirection, $limit, $offset);
         return $result;
     }
@@ -5475,6 +5526,7 @@ SQL;
                     "default" => false,
                     "description" => "Is the category being followed by the current user?",
                 ],
+                "dateFollowed:dt?" => "Date time the user started following the category",
                 "breadcrumbs:a?" => new InstanceValidatorSchema(Breadcrumb::class),
                 "featured:b?" => "Featured category.",
                 "allowedDiscussionTypes:a",
@@ -5577,31 +5629,33 @@ SQL;
      *
      * @param int $categoryID
      */
-    private function prepareForDelete(int $categoryID): void
+    private function prepareForDelete(int $categoryID, bool $deleteContent = true): void
     {
         $this->deletePermissions($categoryID);
-
-        // Delete comments in this category
-        $this->SQL
-            ->from("Comment c")
-            ->join("Discussion d", "c.DiscussionID = d.DiscussionID")
-            ->where("d.CategoryID", $categoryID)
-            ->delete();
 
         // Delete discussions in this category
         $this->SQL->delete("Discussion", ["CategoryID" => $categoryID]);
 
-        // Make inherited permission local permission
-        $this->SQL
-            ->update("Category")
-            ->set("PermissionCategoryID", 0)
-            ->where("PermissionCategoryID", $categoryID)
-            ->where("CategoryID <>", $categoryID)
-            ->put();
+        if ($deleteContent) {
+            // Delete comments in this category
+            $this->SQL
+                ->from("Comment c")
+                ->join("Discussion d", "c.DiscussionID = d.DiscussionID")
+                ->where("d.CategoryID", $categoryID)
+                ->delete();
 
-        // Delete tags
-        $this->SQL->delete("Tag", ["CategoryID" => $categoryID]);
-        $this->SQL->delete("TagDiscussion", ["CategoryID" => $categoryID]);
+            // Make inherited permission local permission
+            $this->SQL
+                ->update("Category")
+                ->set("PermissionCategoryID", 0)
+                ->where("PermissionCategoryID", $categoryID)
+                ->where("CategoryID <>", $categoryID)
+                ->put();
+
+            // Delete tags
+            $this->SQL->delete("Tag", ["CategoryID" => $categoryID]);
+            $this->SQL->delete("TagDiscussion", ["CategoryID" => $categoryID]);
+        }
     }
 
     /**
